+++ mrbutler/src/main/java/com/jetradar/permissions/PermissionsHandler.kt
@@ -19,6 +19,6 @@ package com.jetradar.permissions
 import io.reactivex.Observable
 
 interface PermissionsHandler {
-  fun checkPermissions(vararg permissions: String): Observable<PermissionInfo>
-  fun requestPermissions(vararg permissions: String): Observable<PermissionInfo>
+  fun checkPermissions(vararg permissions: String): Observable<PermissionCheckResult>
+  fun requestPermissions(vararg permissions: String): Observable<PermissionCheckResult>
 }