+++ CHANGELOG.md
@@ -2,6 +2,14 @@ Change Log
 ==========
 
 
+## Version 1.0.1
+
+_2018-04-10_
+
+ * Using sealed class `PermissionCheckResult` instead `PermissionInfo`.
+ * New api `checkSingle` & `requestSingle`.
+
+
 ## Version 1.0.0
 
 _2018-03-06_